+++ src/components/missions/Mission.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
-import { joinMission, leaveMission } from '../redux/missions/missions';
+import { joinMission, leaveMission } from '../../redux/missions/missions';
 import styles from './Mission.module.scss';
 
 const Mission = ({
